Break up props and spread tests
