+++ tests/auto/qtpromise/qpromise/props/props.pro
@@ -1,0 +1,4 @@
+TARGET = tst_qpromise_props
+SOURCES += $$PWD/tst_props.cpp
+
+include(../../qtpromise.pri)
